panel/requirements.txt
# FastAPI - High performance async framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.25

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
apscheduler==3.10.4
aiofiles==23.2.1
jinja2==3.1.3
